implemented User.class, Journal.class, AppDao.class, AppDatabase.java, Util.java, packages db, typeConverters, DateTypeConverter.class implemented LoginActivity.java, user_menu.xml 2023.12.05 update user_menu.xml, activity_main.xml implemented Auction.java, Bid.java implemented MyBroadcastReceiver.java 2023.12.09 added code to set subitem2 visible if current user is admin 2023.12.10 implemented SignUpActivity.java, AdminActivity.java 2023.12.14 implemented ProfileActivity.java, AdminActivity.java 2023.12.15 implemented Item using use cases

+++ app/src/main/java/com/sscire/auctionhouse/ItemActivity.java
@@ -107,6 +107,10 @@ public class ItemActivity extends AppCompatActivity {
                         throw new NumberFormatException ();
                     }
                     String itemName = mItem.getItemName();
+                    int itemUserId = mItem.getUserId();
+                    if(itemUserId != mUserId){
+                        throw new NumberFormatException ();
+                    }
                     mAppDAO.delete(mItem);
                     Toast.makeText(ItemActivity.this,
                             "Item: " + itemId + " " + itemName + " deleted."
@@ -115,7 +119,7 @@ public class ItemActivity extends AppCompatActivity {
                     mItemId.setText("");
                 } catch (NumberFormatException e){
                     Toast.makeText(ItemActivity.this,
-                            "Enter a valid itemId"
+                            "Enter a valid itemId owned by user"
                             , Toast.LENGTH_SHORT).show();
                 }
             }
